Reorg IndexBuilder to separate files
