+++ src.mk
@@ -73,10 +73,12 @@ LIB_SOURCES =                                                   \
   table/format.cc                                               \
   table/full_filter_block.cc                                    \
   table/get_context.cc                                          \
+  table/index_builder.cc                                        \
   table/iterator.cc                                             \
   table/merging_iterator.cc                                     \
   table/meta_blocks.cc                                          \
   table/sst_file_writer.cc                                      \
+  table/partitioned_filter_block.cc                             \
   table/plain_table_builder.cc                                  \
   table/plain_table_factory.cc                                  \
   table/plain_table_index.cc                                    \
