+++ table/full_filter_block.cc
@@ -73,8 +73,8 @@ FullFilterBlockReader::FullFilterBlockReader(
   block_contents_ = std::move(contents);
 }
 
-bool FullFilterBlockReader::KeyMayMatch(const Slice& key,
-    uint64_t block_offset, const bool no_io) {
+bool FullFilterBlockReader::KeyMayMatch(const Slice& key, uint64_t block_offset,
+                                        const bool no_io) {
   assert(block_offset == kNotValid);
   if (!whole_key_filtering_) {
     return true;
@@ -83,7 +83,8 @@ bool FullFilterBlockReader::KeyMayMatch(const Slice& key,
 }
 
 bool FullFilterBlockReader::PrefixMayMatch(const Slice& prefix,
-                                           uint64_t block_offset, const bool no_io) {
+                                           uint64_t block_offset,
+                                           const bool no_io) {
   assert(block_offset == kNotValid);
   if (!prefix_extractor_) {
     return true;
