+++ table/full_filter_block.h
@@ -91,10 +91,11 @@ class FullFilterBlockReader : public FilterBlockReader {
   ~FullFilterBlockReader() {}
 
   virtual bool IsBlockBased() override { return false; }
-  virtual bool KeyMayMatch(const Slice& key,
-                           uint64_t block_offset = kNotValid, const bool no_io = false) override;
+  virtual bool KeyMayMatch(const Slice& key, uint64_t block_offset = kNotValid,
+                           const bool no_io = false) override;
   virtual bool PrefixMayMatch(const Slice& prefix,
-                              uint64_t block_offset = kNotValid, const bool no_io = false) override;
+                              uint64_t block_offset = kNotValid,
+                              const bool no_io = false) override;
   virtual size_t ApproximateMemoryUsage() const override;
 
  protected:
